pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
python_version = 3.10
show_error_context = true
pretty = true
show_column_numbers = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
exclude = ['docs/']

[tool.mypy-gevent]
ignore_missing_imports = true

[tool.mypy-mock]
ignore_missing_imports = true

[tool.mypy-pytest]
ignore_missing_imports = true

[tool.mypy-pytz]
ignore_missing_imports = true

[tool.mypy-volttron]
ignore_missing_imports = true

[tool.poetry]
name = "volttron-message-bus-adapter"
version = "2.0.0rc0"
description = "The VOLTTRON Message Bus Adapter Agent relays data between VOLTTRON and a foreign message bus."
authors = ["The VOLTTRON Development Team <volttron@pnnl.gov>"]
license = "Apache License 2.0"
readme = "README.md"
repository = "https://github.com/eclipse-volttron/volttron-message-bus-adapter"
homepage = "https://github.com/eclipse-volttron/volttron-message-bus-adapter"
keywords = []
packages = [ { include = "protocol_proxy", from = "src" } ]
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: Apache Software License"
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
volttron-core = ">=2.0.0rc0"
protocol-proxy = ">=2.0.0rc0"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
mock = "^4.0.3"
pre-commit = "^2.17.0"
yapf = "^0.32.0"
toml = "^0.10.2"
mypy = "^0.942"
isort = "^5.10.1"

[tool.poetry.group.documentation.dependencies]
Sphinx = "^4.5.0"
sphinx-rtd-theme = "^1.0.0"

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
column_limit = 99
split_before_logical_operator = true

[tool.yapfignore]
ignore_patterns = [
    ".venv/**",
    ".pytest_cache/**",
    "dist/**",
    "docs/**"
]
